pyproject.toml
[project]
name = "llm-tools-clojure"
version = "0.0.1"
description = "Talk to clojure nrepl as an LLM tool"
readme = "README.md"
authors = [{name = "Kawpuh"}]
license = "Apache-2.0"
classifiers = []
requires-python = ">=3.9"
dependencies = [
    "llm",
    "nrepl-python-client"
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project.urls]
Homepage = "https://github.com/kawpuh/llm-tools-clojure"

[project.entry-points.llm]
tools_clojure = "llm_tools_clojure"

[project.optional-dependencies]
test = ["pytest", "llm-echo>=0.3a1"]
